do not add method to api.Client
